+++ shared-lib/impl/BtrfRootBlock.h
@@ -50,15 +50,15 @@ public:
 	virtual int DLLCALLCONV addTemplate(const TemplateGuid& guid, int usedField);
 	virtual const TemplateGuid& DLLCALLCONV getTemplateGuid(int index);
 	virtual int DLLCALLCONV getTemplateUsedField(int index);
-	virtual int DLLCALLCONV getStringNum() { return stringList.size(); }
-	virtual int DLLCALLCONV getTemplateNum() { return templateList.size(); }
+	virtual int DLLCALLCONV getStringNum() { return (int)stringList.size(); }
+	virtual int DLLCALLCONV getTemplateNum() { return (int)templateList.size(); }
 
 	virtual TmlFile * DLLCALLCONV getTmlFile() { return tmlFile; }
 
 	virtual int DLLCALLCONV addBlock(IBtrfBlock *iBlock);
 	virtual BtrfBlock * DLLCALLCONV getBlockByGuid(const TemplateGuid& guid);
 	virtual BtrfBlock * DLLCALLCONV getBlockById(int id) { return blockList.at(id); }
-	virtual int DLLCALLCONV getBlockNum() { return blockList.size(); }
+	virtual int DLLCALLCONV getBlockNum() { return (int)blockList.size(); }
 
 	virtual void DLLCALLCONV dumpToStdout(FILE* fout);
